2007-11-29  Gary Benson  <gbenson@redhat.com>
	* ports/hotspot/src/os_cpu/linux_ppc/vm/assembler_ppc.cpp
	(MacroAssembler::needs_explicit_null_check): Fix broken definition.

+++ ports/hotspot/src/os_cpu/linux_ppc/vm/assembler_linux_ppc.cpp
@@ -26,6 +26,7 @@
 #include "incls/_precompiled.incl"
 #include "incls/_assembler_linux_ppc.cpp.incl"
 
-bool MacroAssembler::needs_explicit_null_check(int offset) {
+bool MacroAssembler::needs_explicit_null_check(intptr_t offset)
+{
   Unimplemented();
 }
